% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_age_risk_points.R
\name{calc_age_points}
\alias{calc_age_points}
\title{calculate age risk factor}
\usage{
calc_age_points(gender, age)
}
\arguments{
\item{gender}{A character}

\item{age}{A number}
}
\value{
framingham age risk scores \code{age} and \code{gender}
}
\description{
\code{calc_age_points} returns age risk factor points given age and gender
}
\details{
\code{calc_age_points} returns TODO: add some details
}
\examples{
calc_age_points(gender='M', age=50)
calc_age_points(gender='F', age=48)
calc_age_points(gender='f', age=35)
}
