% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_bmi_risk_points.R
\name{calc_bmi_points}
\alias{calc_bmi_points}
\title{calculate BMI risk factor}
\usage{
calc_bmi_points(bmi)
}
\arguments{
\item{bmi}{A number}
}
\value{
framingham BMI risk score \code{status} and \code{bmi}
}
\description{
\code{calc_BMI_points} returns BMI risk factor points given BMI status and gender
}
\examples{
calc_bmi_points(bmi)
}
