% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiovascular_10yrs_one.R
\name{calc_card_10_one}
\alias{calc_card_10_one}
\title{Cardivascular Disease 10 years risk for one patient}
\usage{
calc_card_10_one(gender, age, bmi, hdl, cholesterol, sbp,
  is_sbp_under_treatment, smoking_status, diabetes_status)
}
\arguments{
\item{gender}{A character}

\item{age}{A number}

\item{bmi}{A number}

\item{hdl}{A number}

\item{cholesterol}{A number}

\item{sbp}{A number}

\item{is_sbp_under_treatment}{A boolean}

\item{smoking_status}{A boolean}

\item{diabetes_status}{A boolean}
}
\value{
framingham cardivascular 10 years risk score \code{gender} and \code{...}
}
\description{
\code{calc_card_10_one} returns 10 years cardivascular relative risk score for a single patient,
}
\details{
This package uses the widely used framingham risk scoring algorithm used to determine an individual's chances of developing cardiovascular disease.
This package gives an estimate of the probability that a person will develop cardiovascular disease within a specified amount of time (10 years)
}
\examples{
# call frisk function
calc_card_10_one(age=45, gender="F", cholesterol=89,
            hdl=89, sbp=65, is_sbp_under_treatment=TRUE,
            smoking_status=F, diabetes_status=T
)

call frisk simple scoring function using BMI
calc_card_10_one(age=45, gender="M", bmi= 173,
             sbp=54, is_sbp_under_treatment=T,
            smoking_status=F, diabetes_status=T
}
