% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_cholesterol_risk_points.R
\name{calc_cholesterol_points}
\alias{calc_cholesterol_points}
\title{calculate cholesterol risk factor}
\usage{
calc_cholesterol_points(gender, chol)
}
\arguments{
\item{gender}{A character}

\item{cholesterol}{A number}
}
\value{
framingham cholesterol risk score \code{cholesterol} and \code{gender}
}
\description{
\code{calc_cholesterol_points} returns cholesterol risk factor points given cholesterol and gender
}
\examples{
calc_cholesterol_points(gender='f', chol=35)
}
