% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_diabetes_risk_points.R
\name{calc_diabetes_points}
\alias{calc_diabetes_points}
\title{calculate diabetes risk factor}
\usage{
calc_diabetes_points(status, gender)
}
\arguments{
\item{status}{boolean}

\item{gender}{A character}
}
\value{
framingham smoking risk score \code{status} and \code{gender}
}
\description{
\code{calc_diabetes_points} returns diabetes risk factor points given diabetes status and gender
}
\examples{
calc_diabetes_points(status, gender)
}
