% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hdl_risk_points.R
\name{calc_hdl_points}
\alias{calc_hdl_points}
\title{calculate hdl risk factor}
\usage{
calc_hdl_points(hdl)
}
\arguments{
\item{hdl}{A number}
}
\value{
framingham CAD10 hdl risk score \code{hdl}
}
\description{
\code{calc_hdl_points} returns hdl risk factor points given hdl
}
\examples{
calc_hdl_points(60)
calc_hdl_points(30)
}
