% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_sbp_risk_points.R
\name{calc_sbp_points}
\alias{calc_sbp_points}
\title{calculate sbp risk factor}
\usage{
calc_sbp_points(sbp, gender, isTreated = FALSE)
}
\arguments{
\item{sbp}{A number}

\item{gender}{A character}

\item{isTreated}{A boolean}

\item{age}{A number}
}
\value{
framingham hdl risk score \code{sbp} and \code{gender} and \code{isTreated}
}
\description{
\code{calc_sbp_points} returns sbp risk factor points given hdl and gender
}
\examples{
calc_sbp_points(sbp, gender)
}
