% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_smoking_risk_points.R
\name{calc_smoking_points}
\alias{calc_smoking_points}
\title{calculate smoking risk factor}
\usage{
calc_smoking_points(status, gender)
}
\arguments{
\item{status}{boolean}

\item{gender}{A character}

\item{age}{A number}
}
\value{
framingham smoking risk score \code{status} and \code{gender}
}
\description{
\code{calc_smoking_points} returns smoking risk factor points given smoking status and gender
}
\examples{
calc_smoking_points(status, gender)
}
