% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_point_converter.R
\docType{data}
\name{male_risk_csv}
\alias{male_risk_csv}
\title{point converter}
\format{An object of class \code{data.frame} with 45 rows and 4 columns.}
\usage{
male_risk_csv
}
\arguments{
\item{gender}{A character}

\item{points}{A number}

\item{bmi}{A number or NA}
}
\value{
risk and heart_age given \code{points} and \code{gender}  and \code{bmi}
}
\description{
\code{point_converter} returns relative risk and heart age given framingham point and gender
}
\details{
\code{point_converter} returns TODO: add some details
}
\examples{
point_converter(points=1, gender='M', NA)

}
\keyword{datasets}
