% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_point_converter.R
\name{point_converter}
\alias{point_converter}
\title{point converter}
\usage{
point_converter(points, gender, bmi = NA)
}
\arguments{
\item{points}{A number}

\item{gender}{A character}

\item{bmi}{A number or NA}
}
\value{
risk and heart_age given \code{points} and \code{gender}  and \code{bmi}
}
\description{
\code{point_converter} returns relative risk and heart age given framingham point and gender
}
\details{
\code{point_converter} returns TODO: add some details
}
\examples{
point_converter(points=1, gender='M', NA)
}
