% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiovascular_10yrs.R
\name{calc_card_10}
\alias{calc_card_10}
\title{cardivascular mortality risk}
\usage{
calc_card_10(data, ...)
}
\arguments{
\item{data}{A dataframe}

\item{gender}{A character}

\item{age}{A number}

\item{hdl}{A number}

\item{cholesterol}{A number}

\item{sbp}{A number}

\item{is_sbp_under_treatment}{A boolean}

\item{smoking_status}{A boolean}
}
\value{
framingham cardivascular risk scores \code{data} and \code{...}
}
\description{
\code{calc_card_10} returns 10 years cardivascular mortality risk score
}
\details{
This is the main function: which returns framingham cardivascular risk score
}
\examples{
df <- data.frame(age=sample(1:100,10,rep=TRUE),
                 gender=sample(c('M','F'),10,rep=TRUE)
                 )
calc_card_10(df, age="age", gender="gender")
}
