% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiovascular_10yrs_one.R
\name{calc_card_10_one}
\alias{calc_card_10_one}
\title{Cardivascular Disease 10 years risk for one patient}
\usage{
calc_card_10_one(data, ...)
}
\arguments{
\item{data}{A dataframe}

\item{gender}{A character}

\item{age}{A number}

\item{bmi}{A number}

\item{hdl}{A number}

\item{cholesterol}{A number}

\item{sbp}{A number}

\item{is_sbp_under_treatment}{A boolean}

\item{smoking_status}{A boolean}

\item{diabetes_status}{A boolean}
}
\value{
framingham cardivascular 10 years risk score \code{data} and \code{...}
}
\description{
\code{calc_card_10_one} returns 10 years cardivascular relative risk score,
}
\details{
This package uses the widely used framingham risk scoring algorithm used to determine an individual's chances of developing cardiovascular disease.
This package gives an estimate of the probability that a person will develop cardiovascular disease within a specified amount of time (10 years)
}
\examples{
# simulate patients data
df <- data.frame(age=sample(30:70,100,rep=TRUE),
              gender=sample(c("M","F"),100,rep=TRUE),
              bmi=sample(16:48, rep = TRUE),
              hdl=sample(10:100,100,rep=TRUE),
              chl=sample(100:400,100,rep=TRUE),
              sbp=sample(90:200,100,rep=TRUE),
              isSbpTreated=sample(c(TRUE,FALSE),100,rep=TRUE),
              smoking=sample(c(TRUE,FALSE),100,rep=TRUE),
              diabetes=sample(c(TRUE,FALSE),100,rep=TRUE)
)

# call frisk function
calc_card_10_one(df, age="age", gender="gender", cholesterol="chl",
            hdl="hdl", sbp="sbp", is_sbp_under_treatment="isSbpTreated",
            smoking_status="smoking", diabetes_status="diabetes"
)

call frisk simple scoring function using BMI
calc_card_10_one(df, age="age", gender="gender", bmi= "BMI",
             sbp="sbp", is_sbp_under_treatment="isSbpTreated",
            smoking_status="smoking", diabetes_status="diabetes"
}
