% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hdl_risk_points.R
\name{calc_hdl_points}
\alias{calc_hdl_points}
\title{calculate hdl risk factor}
\usage{
calc_hdl_points(hdl, gender)
}
\arguments{
\item{hdl}{A number}

\item{gender}{A character}
}
\value{
framingham hdl risk score \code{hdl} and \code{gender}
}
\description{
\code{calc_hdl_points} returns hdl risk factor points given hdl and gender
}
\examples{
calc_hdl_points(hdl, gender)
}
